requirements.txt
python-dotenv
rich
langchain-core
serpapi
